Move binary expression classes to their own files
